RF - merge floating, casting, refine clip tests
floating, casting into one file.
When testing for int_clippers and casting, allow for possibility that
the overflow generates the integer maximum / minimum rather than a very
negative value (in the max overflow case)
