+++ nibabel/tests/test_casting.py
@@ -3,7 +3,7 @@
 
 import numpy as np
 
-from ..casting import nice_round, int_clippers, CastingError
+from ..casting import (float_to_int, int_clippers, CastingError, int_to_float)
 
 from numpy.testing import (assert_array_almost_equal, assert_array_equal)
 
@@ -14,14 +14,34 @@ def test_int_clippers():
     for ft in np.sctypes['float']:
         for it in np.sctypes['int'] + np.sctypes['uint']:
             # Test that going a bit above or below the calculated min and max
-            # either generates the same number when cast, or something smaller
-            # (as a result of overflow)
+            # either generates the same number when cast, or the max int value
+            # (if this system generates that) or something smaller (because of
+            # overflow)
             mn, mx = int_clippers(ft, it)
             ovs = ft(mx) + np.arange(2048, dtype=ft)
             # Float16 can overflow to inf
             bit_bigger = ovs[np.isfinite(ovs)].astype(it)
             casted_mx = ft(mx).astype(it)
-            assert_true(np.all((bit_bigger <= casted_mx)))
+            imax = int(np.iinfo(it).max)
+            thresh_overflow = False
+            if casted_mx != imax:
+                # The int_clippers have told us that they believe the imax does
+                # not have an exact representation.
+                fimax = int_to_float(imax, ft)
+                if np.isfinite(fimax):
+                    assert_true(int(fimax) != imax)
+                # Therefore the imax, cast back to float, and to integer, will
+                # overflow. If it overflows to the imax, we need to allow for
+                # that possibility in the testing of our overflowed values
+                imax_roundtrip = fimax.astype(it)
+                if imax_roundtrip == imax:
+                    thresh_overflow = True
+            if thresh_overflow:
+                assert_true(np.all(
+                    (bit_bigger == casted_mx) |
+                    (bit_bigger == imax)))
+            else:
+                assert_true(np.all((bit_bigger <= casted_mx)))
             if it in np.sctypes['uint']:
                 assert_equal(mn, 0)
                 continue
@@ -30,7 +50,25 @@ def test_int_clippers():
             # Float16 can overflow to inf
             bit_smaller = ovs[np.isfinite(ovs)].astype(it)
             casted_mn = ft(mn).astype(it)
-            assert_true(np.all(bit_smaller >= casted_mn))
+            imin = int(np.iinfo(it).min)
+            if casted_mn != imin:
+                # The int_clippers have told us that they believe the imin does
+                # not have an exact representation.
+                fimin = int_to_float(imin, ft)
+                if np.isfinite(fimin):
+                    assert_true(int(fimin) != imin)
+                # Therefore the imin, cast back to float, and to integer, will
+                # overflow. If it overflows to the imin, we need to allow for
+                # that possibility in the testing of our overflowed values
+                imin_roundtrip = fimin.astype(it)
+                if imin_roundtrip == imin:
+                    thresh_overflow = True
+            if thresh_overflow:
+                assert_true(np.all(
+                    (bit_smaller == casted_mn) |
+                    (bit_smaller == imin)))
+            else:
+                assert_true(np.all((bit_smaller >= casted_mn)))
 
 
 def test_casting():
@@ -38,12 +76,14 @@ def test_casting():
         for it in np.sctypes['int'] + np.sctypes['uint']:
             ii = np.iinfo(it)
             arr = [ii.min-1, ii.max+1, -np.inf, np.inf, np.nan, 0.2, 10.6]
-            farr = np.array(arr, dtype=ft)
+            farr_orig = np.array(arr, dtype=ft)
+            # We're later going to test if we modify this array
+            farr = farr_orig.copy()
             mn, mx = int_clippers(ft, it)
-            iarr = nice_round(farr, it)
+            iarr = float_to_int(farr, it)
             exp_arr = np.array([mn, mx, mn, mx, 0, 0, 11])
             assert_array_equal(iarr, exp_arr)
-            iarr = nice_round(farr, it, infmax=True)
+            iarr = float_to_int(farr, it, infmax=True)
             # Float16 can overflow to infs
             if farr[0] == -np.inf:
                 exp_arr[0] = ii.min
@@ -60,10 +100,12 @@ def test_casting():
                 exp_arr[3] = ii.max
             assert_array_equal(iarr, exp_arr)
             # Confirm input array is not modified
-            assert_array_equal(farr, np.array(arr, dtype=ft))
+            nans = np.isnan(farr)
+            assert_array_equal(nans, np.isnan(farr_orig))
+            assert_array_equal(farr[nans==False], farr_orig[nans==False])
     # Test scalars work and return scalars
-    assert_array_equal(nice_round(np.float32(0), np.int16), [0])
+    assert_array_equal(float_to_int(np.float32(0), np.int16), [0])
     # Test scalar nan OK
-    assert_array_equal(nice_round(np.nan, np.int16), [0])
+    assert_array_equal(float_to_int(np.nan, np.int16), [0])
     # Test nans give error if not nan2zero
-    assert_raises(CastingError, nice_round, np.nan, np.int16, False)
+    assert_raises(CastingError, float_to_int, np.nan, np.int16, False)
